#61 [FEAT] : Home category click event UI 구현
